Refactor PaymentService
Now the whole royalty payment calculus is done in the service, and there are no utility method in RoyaltyPayment which now contains only final fields. Using stream API usually means apply a "functionalish" approach, and this at least means that funcions should be pure and passed instances should be immutable, otherwise the worst of imperative and function will apply.