Refactor UtilsJSON
Static instances just limit the GC to release the memory of those instances, so it is a bad idea. Now, the ObjectMapper is used during loading, but then in completely released from the heap space. Also, creating those simple wrappers (Episodes and Studios) parsing is much easier and type safe.